# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customers": {
      "post": {
        "operationId": "Customers_create",
        "summary": "Create a new customer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "Customers_findAll",
        "summary": "Get all customers with pagination and search",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by full name or phone number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/customers/lookup": {
      "post": {
        "operationId": "Customers_findByPhone",
        "summary": "Lookup the minimum customer profile needed at POS",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "operationId": "Customers_findOne",
        "summary": "Get customer by ID with recent orders and point history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "Customers_update",
        "summary": "Update customer information",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/customers/{id}/points": {
      "get": {
        "operationId": "Customers_getPointHistory",
        "summary": "Get point history for a customer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by full name or phone number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/api/v1/hr/announcements": {
      "post": {
        "operationId": "Announcement_create",
        "summary": "Create an announcement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnouncementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "Announcement_findAll",
        "summary": "Get all active announcements",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnouncementResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/hr/announcements/{id}": {
      "get": {
        "operationId": "Announcement_findOne",
        "summary": "Get an announcement by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "Announcement_remove",
        "summary": "Delete an announcement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "HR"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/finance/cash-flows": {
      "post": {
        "operationId": "Finance_create",
        "summary": "Create a cash flow entry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCashFlowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Finance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "Finance_findAll",
        "summary": "Get all cash flows with filters and pagination",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "description": "Filter by Branch ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by CashFlowType",
            "schema": {
              "enum": [
                "INCOME",
                "EXPENSE"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date filter",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date filter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CashFlowResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Finance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/finance/summary": {
      "get": {
        "operationId": "Finance_getSummary",
        "summary": "Get financial summary",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowSummaryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Finance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TeaP ERP/POS API",
    "description": "API documentation for TeaP — Bubble Tea Chain ERP/POS Management System. Supports multi-branch operations, POS, inventory management, HR, and finance.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT access token",
        "in": "header"
      }
    },
    "schemas": {
...
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Full name of the customer"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the customer (VN format)"
          },
          "email": {
            "type": "string",
            "description": "Email address of the customer"
          }
        },
        "required": [
          "fullName",
          "phone"
        ]
      },
      "LookupCustomerDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "0901234567"
          }
        },
        "required": [
          "phone"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Full name of the customer"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the customer (VN format)"
          },
          "email": {
            "type": "string",
            "description": "Email address of the customer"
          }
        }
      },
...
      "CreateAnnouncementDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
      "AnnouncementResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "authorId": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "authorId",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CashFlowDateSummaryDto": {
        "type": "object",
        "properties": {
          "income": {
            "type": "integer",
            "description": "Whole VND"
          },
          "expense": {
            "type": "integer",
            "description": "Whole VND"
          }
        },
        "required": [
          "income",
          "expense"
        ]
      },
      "CreateCashFlowDto": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "string",
            "description": "Branch ID"
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ],
            "description": "Type of cash flow"
          },
          "amount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9999999999,
            "description": "Amount in whole VND",
            "example": 50000
          },
          "description": {
            "type": "string",
            "description": "Description of the cash flow"
          },
          "refType": {
            "type": "string",
            "description": "Reference Type (e.g. ORDER, SALARY)"
          },
          "refId": {
            "type": "string",
            "description": "Reference ID"
          }
        },
        "required": [
          "branchId",
          "type",
          "amount",
          "description"
        ]
      },
      "CashFlowBranchResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "CashFlowResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "amount": {
            "type": "integer",
            "description": "Whole VND",
            "example": 150000
          },
          "description": {
            "type": "string"
          },
          "refType": {
            "type": "string",
            "nullable": true
          },
          "refId": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "branch": {
            "$ref": "#/components/schemas/CashFlowBranchResponseDto"
          }
        },
        "required": [
          "id",
          "branchId",
          "type",
          "amount",
          "description",
          "createdBy",
          "createdAt"
        ]
      },
      "CashFlowSummaryResponseDto": {
        "type": "object",
        "properties": {
          "totalIncome": {
            "type": "integer",
            "description": "Whole VND"
          },
          "totalExpense": {
            "type": "integer",
            "description": "Whole VND"
          },
          "netCashFlow": {
            "type": "integer",
            "description": "Whole VND"
          },
          "groupedByDate": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/CashFlowDateSummaryDto"
            }
          }
        },
        "required": [
          "totalIncome",
          "totalExpense",
          "netCashFlow",
          "groupedByDate"
        ]
      }
    }
  }
}